refactor: isolate parser error
